requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
django-anymail==8.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
localStoragePy==0.2.3
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
dj-database-url==1.0.0
psycopg2-binary==2.9.5